requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Experiment Tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Feature Engineering
nltk>=3.8
spacy>=3.5.0
textblob>=0.17.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0
isort>=5.12.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
